+++ app/Modules/Company/Infrastructure/Controllers/CompanyController.php
@@ -9,10 +9,12 @@ use App\Modules\Company\Application\Resources\CompanyResource;
 use App\Modules\Company\Application\Resources\CompanyCollection;
 use App\Modules\Company\Application\Resources\BranchCollection;
 use App\Modules\Company\Domain\Models\Company;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 
 class CompanyController extends Controller
 {
+    use AuthorizesRequests;
     protected $repository;
 
     public function __construct(CompanyRepository $repository)
@@ -20,14 +22,16 @@ class CompanyController extends Controller
         $this->repository = $repository;
     }
 
-    public function index(Request $request) 
+    public function index(Request $request)
     {
+        $this->authorize('viewAny', Company::class);
         $companies = $this->repository->all($request->all(), $request->input('perPage'));
         return new CompanyCollection($companies);
     }
 
     public function store(CompanyRequest $request)
     {
+        $this->authorize('create', Company::class);
         $company = $this->repository->create($request->validated());
         return new CompanyResource($company);
     }
@@ -36,19 +40,23 @@ class CompanyController extends Controller
     {
         $includes = $request->query('include', '');
         $includes = $includes ? explode(',', $includes) : [];
-
         $company = $this->repository->find($id, $includes);
+        $this->authorize('view', $company);
         return new CompanyResource($company);
     }
 
     public function update(CompanyRequest $request, int $id)
     {
+        $company = $this->repository->find($id);
+        $this->authorize('update', $company);
         $company = $this->repository->update($request->validated(), $id);
         return new CompanyResource($company);
     }
 
     public function destroy(int $id)
     {
+        $company = $this->repository->find($id);
+        $this->authorize('delete', $company);
         $this->repository->delete($id);
         return response()->json(['message' => 'Company deleted']);
     }
